+++ tests/test_run_length.py
@@ -255,7 +255,7 @@ class TestRunsWithDates:
             ("07-01", 210, 70),
             ("07-01", 190, 50),
             ("04-01", 150, np.NaN),  # date falls early
-            ("11-01", 150, 164),  # date ends late
+            ("11-01", 150, 165),  # date ends late
             (None, 150, 10),  # no date, real length
         ],
     )
@@ -284,7 +284,7 @@ class TestRunsWithDates:
             ("dayofyear", "07-01", 210, 211),
             (False, "07-01", 190, 190),
             ("dayofyear", "04-01", 150, np.NaN),  # date falls early
-            ("dayofyear", "11-01", 150, 305),  # date ends late
+            ("dayofyear", "11-01", 150, 306),  # date ends late
         ],
     )
     @pytest.mark.parametrize("use_dask", [True, False])
@@ -307,7 +307,7 @@ class TestRunsWithDates:
             ("dayofyear", "07-01", 210, 211),
             (False, "07-01", 190, 190),
             ("dayofyear", "04-01", False, np.NaN),  # no run
-            ("dayofyear", "11-01", 150, 305),  # run already started
+            ("dayofyear", "11-01", 150, 306),  # run already started
         ],
     )
     @pytest.mark.parametrize("use_dask", [True, False])
@@ -332,8 +332,8 @@ class TestRunsWithDates:
     @pytest.mark.parametrize(
         "coord,date,end,expected",
         [
-            ("dayofyear", "07-01", 210, 182),
-            (False, "07-01", 190, 181),
+            ("dayofyear", "07-01", 210, 183),
+            (False, "07-01", 190, 182),
             ("dayofyear", "04-01", 150, np.NaN),  # date falls early
             ("dayofyear", "11-01", 150, 150),  # date ends late
         ],
@@ -372,6 +372,55 @@ class TestRunsWithDates:
 
         out = runs.resample(time="MS").map(func, window=1, date="07-01", dim="time")
         assert out.isnull().all()
+
+    @pytest.mark.parametrize(
+        "calendar,expected",
+        [("standard", [61, 60]), ("365_day", [60, 60]), ("366_day", [61, 61])],
+    )
+    def test_run_with_dates_different_calendars(self, calendar, expected):
+        time = xr.cftime_range(
+            "2004-01-01", end="2005-12-31", freq="D", calendar=calendar
+        )
+        tas = np.zeros(time.size)
+        start = np.where((time.day == 1) & (time.month == 3))[0]
+        tas[start[0] : start[0] + 250] = 5
+        tas[start[1] : start[1] + 250] = 5
+        tas = xr.DataArray(tas, coords={"time": time}, dims=("time",))
+        out = (
+            (tas > 0)
+            .resample(time="AS-MAR")
+            .map(rl.first_run_after_date, date="03-01", window=2)
+        )
+        np.testing.assert_array_equal(out.values[1:], expected)
+
+        out = (
+            (tas > 0)
+            .resample(time="AS-MAR")
+            .map(rl.season_length, date="03-02", window=2)
+        )
+        np.testing.assert_array_equal(out.values[1:], [250, 250])
+
+        out = (
+            (tas > 0)
+            .resample(time="AS-MAR")
+            .map(rl.run_end_after_date, date="03-03", window=2)
+        )
+        np.testing.assert_array_equal(out.values[1:], np.array(expected) + 250)
+
+        out = (
+            (tas > 0)
+            .resample(time="AS-MAR")
+            .map(rl.last_run_before_date, date="03-02", window=2)
+        )
+        np.testing.assert_array_equal(out.values[1:], np.array(expected) + 1)
+
+    @pytest.mark.parametrize(
+        "func", [rl.first_run_after_date, rl.season_length, rl.run_end_after_date]
+    )
+    def test_too_many_dates(self, func, tas_series):
+        tas = tas_series(np.zeros(730), start="2000-01-01")
+        with pytest.raises(ValueError, match="More than 1 instance of date"):
+            func((tas == 0), date="03-01", window=5)
 
 
 @pytest.mark.parametrize("use_dask", [True, False])